Hieno säätöä

+++ fishlist/src/main/java/data/Point.java
@@ -18,7 +18,7 @@ public class Point {
 		this.id = id;
 	}
 
-	public float getPoint() {
+	public int getPoint() {
 		return point;
 	}
 	public void setPoint(int point) {
